requirements.txt
flash
pdfplumber
fpdf2
pandas